python_service/pyproject.toml
[tool.poetry]
name = "visionhub-service"
version = "0.1.0"
description = "VisionHub - Industrial Visual Inspection Unified Platform AI Service"
authors = ["LuminaSmart"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pyyaml = "^6.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
numpy = "^1.26"
Pillow = "^10.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
pytest = "^8.0"

[tool.poetry.scripts]
visionhub = "app.main:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
